+++ service/gluedatabrew/doc.go
@@ -3,7 +3,7 @@
 // Package gluedatabrew provides the client and types for making API
 // requests to AWS Glue DataBrew.
 //
-// AWS Glue DataBrew is a visual, cloud-scale data-preparation service. DataBrew
+// Glue DataBrew is a visual, cloud-scale data-preparation service. DataBrew
 // simplifies data preparation tasks, targeting data issues that are hard to
 // spot and time-consuming to fix. DataBrew empowers users of all technical
 // levels to visualize the data and perform one-click data transformations,