+++ service/kendra/api.go
@@ -4742,21 +4742,23 @@ type CapacityUnitsConfiguration struct {
 	// The amount of extra query capacity for an index and GetQuerySuggestions (https://docs.aws.amazon.com/kendra/latest/dg/API_GetQuerySuggestions.html)
 	// capacity.
 	//
-	// A single extra capacity unit for an index provides 0.5 queries per second
-	// or approximately 40,000 queries per day.
+	// A single extra capacity unit for an index provides 0.1 queries per second
+	// or approximately 8,000 queries per day.
 	//
-	// GetQuerySuggestions capacity is 5 times the provisioned query capacity for
-	// an index. For example, the base capacity for an index is 0.5 queries per
-	// second, so GetQuerySuggestions capacity is 2.5 calls per second. If adding
-	// another 0.5 queries per second to total 1 queries per second for an index,
-	// the GetQuerySuggestions capacity is 5 calls per second.
+	// GetQuerySuggestions capacity is five times the provisioned query capacity
+	// for an index, or the base capacity of 2.5 calls per second, whichever is
+	// higher. For example, the base capacity for an index is 0.1 queries per second,
+	// and GetQuerySuggestions capacity has a base of 2.5 calls per second. If you
+	// add another 0.1 queries per second to total 0.2 queries per second for an
+	// index, the GetQuerySuggestions capacity is 2.5 calls per second (higher than
+	// five times 0.2 queries per second).
 	//
 	// QueryCapacityUnits is a required field
 	QueryCapacityUnits *int64 `type:"integer" required:"true"`
 
 	// The amount of extra storage capacity for an index. A single capacity unit
-	// for an index provides 150 GB of storage space or 500,000 documents, whichever
-	// is reached first.
+	// provides 30 GB of storage space or 100,000 documents, whichever is reached
+	// first.
 	//
 	// StorageCapacityUnits is a required field
 	StorageCapacityUnits *int64 `type:"integer" required:"true"`
@@ -6300,6 +6302,9 @@ type CreateIndexInput struct {
 	//
 	// The Edition parameter is optional. If you don't supply a value, the default
 	// is ENTERPRISE_EDITION.
+	//
+	// For more information on quota limits for enterprise and developer editions,
+	// see Quotas (https://docs.aws.amazon.com/kendra/latest/dg/quotas.html).
 	Edition *string `type:"string" enum:"IndexEdition"`
 
 	// The name for the new index.
@@ -13786,9 +13791,8 @@ type SharePointConfiguration struct {
 	InclusionPatterns []*string `type:"list"`
 
 	// The Amazon Resource Name (ARN) of credentials stored in AWS Secrets Manager.
-	// The credentials should be a user/password pair. If you use SharePoint Sever,
-	// you also need to provide the sever domain name as part of the credentials.
-	// For more information, see Using a Microsoft SharePoint Data Source (https://docs.aws.amazon.com/kendra/latest/dg/data-source-sharepoint.html).
+	// The credentials should be a user/password pair. For more information, see
+	// Using a Microsoft SharePoint Data Source (https://docs.aws.amazon.com/kendra/latest/dg/data-source-sharepoint.html).
 	// For more information about AWS Secrets Manager, see What Is AWS Secrets Manager
 	// (https://docs.aws.amazon.com/secretsmanager/latest/userguide/intro.html)
 	// in the AWS Secrets Manager user guide.